requirements.txt
contextlib2==0.4.0
funcsigs==0.4
mock==1.3.0
pbr==1.5.0
Pillow==2.9.0
py==1.4.30
pytest==2.7.2
PyYAML==3.11
requests==2.7.0
six==1.9.0
vcrpy==1.7.3
wheel==0.24.0
wrapt==1.10.5
